refactor bridge modal components
