+++ src/services/EcoBridge/EcoBridge.hooks.ts
@@ -27,7 +27,7 @@ import { useBytes32TokenContract, useTokenContract, useWrappingToken } from '../
 import { isAddress } from '../../utils'
 import { currencyId } from '../../utils/currencyId'
 
-import { BridgeTxsFilter } from './EcoBridge.types'
+import { BridgeModalState, BridgeModalStatus, BridgeTxsFilter } from './EcoBridge.types'
 import { WrappedTokenInfo } from '../../state/lists/wrapped-token-info'
 
 export const useBridgeSupportedTokens = () => {
@@ -315,5 +315,62 @@ export const useBridgeCollectHandlers = () => {
     collectableTx,
     setCollectableTx,
     collectableCurrency: collectableCurrency ?? nativeCurrency
+  }
+}
+
+export const useBridgeModal = (): {
+  modalData: BridgeModalState
+  setModalState: (status: BridgeModalStatus, error?: string) => void
+  setModalData: ({
+    symbol,
+    typedValue,
+    fromChainId,
+    toChainId
+  }: Pick<BridgeModalState, 'symbol' | 'typedValue' | 'fromChainId' | 'toChainId'>) => void
+} => {
+  const { fromChainId, status, symbol, toChainId, typedValue, error, disclaimerText } = useSelector(
+    (state: AppState) => state.ecoBridge.UI.modal
+  )
+
+  const dispatch = useDispatch()
+
+  const setModalState = useCallback(
+    (status: BridgeModalStatus, error?: string) => {
+      dispatch(ecoBridgeUIActions.setBridgeModalStatus({ status, error }))
+    },
+    [dispatch]
+  )
+
+  const setModalData = useCallback(
+    ({
+      symbol,
+      typedValue,
+      fromChainId,
+      toChainId
+    }: Pick<BridgeModalState, 'symbol' | 'typedValue' | 'fromChainId' | 'toChainId'>) => {
+      dispatch(
+        ecoBridgeUIActions.setBridgeModalData({
+          symbol: symbol ?? '',
+          typedValue,
+          fromChainId,
+          toChainId
+        })
+      )
+    },
+    [dispatch]
+  )
+
+  return {
+    modalData: {
+      status,
+      symbol,
+      typedValue,
+      fromChainId,
+      toChainId,
+      error,
+      disclaimerText
+    },
+    setModalState,
+    setModalData
   }
 }
